requirements.txt
# ── Core ─────────────────────────────────────────────────────
httpx>=0.27.0
pydantic>=2.0.0

# ── Database ─────────────────────────────────────────────────
psycopg2-binary>=2.9.9

# ── ETL Scheduler ────────────────────────────────────────────
APScheduler>=3.10.4

# ── Analytics ────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0

# ── Dashboard ────────────────────────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0

# ── REST API ─────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# ── Caching ──────────────────────────────────────────────────
redis>=5.0.4

# ── Dev / Test ────────────────────────────────────────────────
pytest>=8.2.0
pytest-cov>=5.0.0
